Bugfixes for Canvas
Corrected the drawPixel function to correctly place pixels now (there
was an issue with the if statement guard used to not leave the outer
bounds of the pxel buffer).

Corrected the destroyCanvas function. I completely forgot taht functions
basically copy a pointer if you pass that in. So my attempts to set the
canvas pointer to NULL were obviously not working at all. I've corrected
that by passing in a pointer to the pointer and performing operations on
that. Im not sure this is the best way but it will do for now.