Implemented Canvas to PPM
Fixed up and created the function that will grab the contents of a ppm
file and return a pointer to a char buffer that holds the contents of
what will be a ppm file. I need to get rid of some magic constants in
there and maybe check that the allocation of the buffer is entirely
correct, other than that it seems to be working ok!