Updated Canvas -> PPM Conversion
Now accounting for max ppm line length in the canvas to ppm conversion.
This is probably not needed in modern ppm viewers but it was fun to do
anyways. Basically just kept count of the line width and used the new
getRGBByIndex function to add each R,G and B value seperately to the ppm
buffer.