Removed Unnecessary Check For End of Line
Realised that there was no need to have the check for a new line within
the width loop, so I removed that and have it ont he outside in the
height loop. I also adjusted the if statement checking whether a space
char should be added to simply check if the previous char is a newline
(and add a space char if this is not the case).