Created Tests and Functionality for Saving PPM File.
Generic function that takes a const char[] file_path and a char*
pointing to the desired contents of said file and saves the contents to
the file. Currently returns true on success and false on error, probably
needs a more 'C' way of doing this.

Also made a test for this that saves a test_ppm into /tmp.