Started Work On Matrix
I've added in a Matrix struct, a constructor for said struct and a
function for accessing a cell in a given matrice. The accessor function
actually returns a pointer to a given cell which makes it essentially a
getter and a setter.I may change this eventually to two seperate
functions, depending on how unwieldy it gets.

TODO: Add tests (going a bit backwards here).