Reworked Determinant and Cofactor Functionality
Created a Mat3 class to hold submatrices, along with create/destruction
functions for it.
Made a cofactor public function to clean up repeated code in the
determinant and inverse computations.
Got rid of usuage of the result param float[3][3] as a result of this.
Just makes things a bit tidier as it was not really fitting the fomat of
all the other code. TODO tests for all this stuff!